+++ srunner/scenarios/signalized_junction_right_turn.py
@@ -40,6 +40,7 @@ class SignalizedJunctionRightTurn(BasicScenario):
         self._brake_value = 0.5
         self._ego_distance = 40
         self._traffic_light = None
+        self._other_actor_transform = None
 
         super(SignalizedJunctionRightTurn, self).__init__("HeroActorTurningRightAtSignalizedJunction",
                                                           ego_vehicle,
@@ -61,6 +62,19 @@ class SignalizedJunctionRightTurn(BasicScenario):
             sys.exit(-1)
         traffic_light_other.set_state(carla.TrafficLightState.Green)
         traffic_light_other.set_green_time(self.timeout)
+
+    def _initialize_actors(self, config):
+        """
+        Custom initialization
+        """
+        self._other_actor_transform = config.other_actors[0].transform
+        first_vehicle_transform = carla.Transform(
+            carla.Location(config.other_actors[0].transform.location.x,
+                           config.other_actors[0].transform.location.y,
+                           config.other_actors[0].transform.location.z - 5),
+            config.other_actors[0].transform.rotation)
+        first_vehicle = CarlaActorPool.request_new_actor(config.other_actors[0].model, first_vehicle_transform)
+        self.other_actors.append(first_vehicle)
 
     def _create_behavior(self):
         """
@@ -109,10 +123,12 @@ class SignalizedJunctionRightTurn(BasicScenario):
 
         # Behavior tree
         sequence = py_trees.composites.Sequence()
+        sequence.add_child(ActorTransformSetter(self.other_actors[0], self._other_actor_transform))
         sequence.add_child(sync_arrival_parallel)
         sequence.add_child(move_actor_parallel)
         sequence.add_child(stop)
         sequence.add_child(end_condition)
+        sequence.add_child(ActorDestroy(self.other_actors[0]))
 
         return sequence